ProgressStep: Chaning signature of getValue and setValue to double (instead of int) to allow more smooth progression steps

+++ src/main/java/api/ProgressStep.java
@@ -1,9 +1,9 @@
 package api;
 
 public interface ProgressStep {
-    void setProgressPercentage(int value);
+    void setProgressPercentage(double value);
 
     void setText(String text);
 
-    int getProgressPercentage();
+    double getProgressPercentage();
 }
